Move package withheld tools behind admin tools API.
